fix: sync C++ helpers (Mmul/Blas/Cutlass/attention/ShapeBuffer) from ag_new_release_updates_2
