+++ libnd4j/include/helpers/cpu/MmulHelper.cpp
@@ -27,6 +27,8 @@
 #include <execution/Threads.h>
 #include <helpers/BlasHelper.h>
 #include <helpers/ShapeUtils.h>
+#include <helpers/MklBlasHelper.h>
+#include <system/Environment.h>
 #include <system/env_functions.h>
 #include <system/openmp_pragmas.h>
 namespace sd {
@@ -431,7 +433,7 @@ NDArray* MmulHelper::mmulMxM( NDArray* A,  NDArray* B, NDArray* C, const double 
     // which would reinterpret B/C buffers with wrong element size.
     if (!ABC) {
       // Promote to A's type — cast B and C if they differ.
-      NDArray* castB = (bType != aType) ? new NDArray(B->cast(aType)) : nullptr;
+      NDArray* castB = (bType != aType) ? B->cast(aType) : nullptr;
       std::vector<LongType> cShape = {M, N};
       NDArray* castC = (cType != aType) ? new NDArray(NDArray('c', cShape, aType, C->getContext())) : nullptr;
       NDArray* effB = (castB != nullptr) ? castB : const_cast<NDArray*>(B);
@@ -939,7 +941,107 @@ NDArray* MmulHelper::mmulNxN( NDArray* A,  NDArray* B, NDArray* C, const double 
   return C;
 }
 
+//////////////////////////////////////////////////////////////////////////
+// Strided batch GEMM - most efficient for contiguous batched data
+// Uses cblas_sgemm_batch_strided/cblas_dgemm_batch_strided (MKL on CPU)
+// Only handles simple row-major contiguous case for reliability
+// Non-standard layouts fall through to tryBlasBatched/tryBlasPerBatch
+//////////////////////////////////////////////////////////////////////////
+bool MmulHelper::tryBlasStridedBatched(NDArray* A, NDArray* B, NDArray* C,
+                                        double alpha, double beta) {
+#if defined(HAVE_MKL)
+  if (!Environment::getInstance().isEnableBlas()) {
+    return false;
+  }
 
+  // Check if strided batch GEMM is available
+  if (!BlasHelper::getInstance().hasStridedBatchGEMM()) {
+    return false;
+  }
 
+  const int aRank = A->rankOf();
+  const int bRank = B->rankOf();
+  const int cRank = C->rankOf();
+
+  // Only handle 3D tensors for simplicity (most common case)
+  if (aRank != 3 || bRank != 3 || cRank != 3) {
+    return false;
+  }
+
+  const auto xType = A->dataType();
+  const auto yType = B->dataType();
+  const auto zType = C->dataType();
+
+  // Types must match
+  if (xType != yType || yType != zType) {
+    return false;
+  }
+
+  // Only float and double supported for strided batch
+  if (xType != DataType::FLOAT32 && xType != DataType::DOUBLE) {
+    return false;
+  }
+
+  // Validate buffers
+  if (A->buffer() == nullptr || B->buffer() == nullptr || C->buffer() == nullptr) {
+    return false;
+  }
+
+  // Get dimensions: A[bS,M,K] @ B[bS,K,N] = C[bS,M,N]
+  const int batchSize = static_cast<int>(A->sizeAt(0));
+  const int M = static_cast<int>(A->sizeAt(1));
+  const int K = static_cast<int>(A->sizeAt(2));
+  const int K2 = static_cast<int>(B->sizeAt(1));
+  const int N = static_cast<int>(B->sizeAt(2));
+
+  // Validate dimensions
+  if (M <= 0 || K <= 0 || N <= 0 || batchSize <= 0 || K != K2) {
+    return false;
+  }
+
+  // Require strict row-major contiguous layout for all matrices
+  const LongType expectedStrideA = static_cast<LongType>(M) * K;
+  const LongType expectedStrideB = static_cast<LongType>(K) * N;
+  const LongType expectedStrideC = static_cast<LongType>(M) * N;
+
+  const bool aValid = (A->strideAt(2) == 1) && (A->strideAt(1) == K) && (A->strideAt(0) == expectedStrideA);
+  const bool bValid = (B->strideAt(2) == 1) && (B->strideAt(1) == N) && (B->strideAt(0) == expectedStrideB);
+  const bool cValid = (C->strideAt(2) == 1) && (C->strideAt(1) == N) && (C->strideAt(0) == expectedStrideC);
+
+  if (!aValid || !bValid || !cValid) {
+    return false;
+  }
+
+  // Use MKL's strided batch GEMM
+  auto blasLock = BlasHelper::getInstance().lockBlas();
+
+  if (xType == DataType::FLOAT32) {
+    sd::mkl::sgemmBatchStrided(
+        CblasRowMajor, CblasNoTrans, CblasNoTrans,
+        M, N, K,
+        static_cast<float>(alpha),
+        A->bufferAsT<float>(), K, static_cast<MKL_INT>(expectedStrideA),
+        B->bufferAsT<float>(), N, static_cast<MKL_INT>(expectedStrideB),
+        static_cast<float>(beta),
+        C->bufferAsT<float>(), N, static_cast<MKL_INT>(expectedStrideC),
+        batchSize);
+  } else {
+    sd::mkl::dgemmBatchStrided(
+        CblasRowMajor, CblasNoTrans, CblasNoTrans,
+        M, N, K,
+        alpha,
+        A->bufferAsT<double>(), K, static_cast<MKL_INT>(expectedStrideA),
+        B->bufferAsT<double>(), N, static_cast<MKL_INT>(expectedStrideB),
+        beta,
+        C->bufferAsT<double>(), N, static_cast<MKL_INT>(expectedStrideC),
+        batchSize);
+  }
+
+  return true;
+#else
+  // MKL not available - strided batch not supported here
+  return false;
+#endif
+}
 
 }  // namespace sd
