+++ libnd4j/include/helpers/cpu/PointersManager.cpp
@@ -20,8 +20,6 @@
 // @author Yurii Shyrma (iuriish@yahoo.com), created on 06.02.2019
 //
 
-#ifndef SD_CUDA
-
 #include <helpers/PointersManager.h>
 #include <helpers/logger.h>
 #include <memory/Workspace.h>
@@ -54,5 +52,3 @@ PointersManager::~PointersManager() {
 }
 
 }  // namespace sd
-
-#endif
