+++ libnd4j/include/helpers/cuda/ConstantHelper.cu
@@ -45,7 +45,7 @@ namespace sd {
 namespace {
 SD_INLINE cudaStream_t captureSafeStreamOrDefault() {
   if (tl_graphExecutionActive && tl_graphCaptureStream != nullptr) {
-    return tl_graphCaptureStream;
+    return reinterpret_cast<cudaStream_t>(tl_graphCaptureStream);
   }
   auto* streamPtr = LaunchContext::defaultContext()->getCudaStream();
   return (streamPtr != nullptr) ? *streamPtr : nullptr;
